\name{beta10_3}
\alias{beta10_3}
\docType{data}
\title{A randomly generated data}
\description{beta10_3 is a randomly generated data from beta
distribution with parameters of 10 and 3.}
\usage{data(beta10_3)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\references{http://www.sciencedirect.com/science/article/B6V8V-4NHM520-1/2/a21e0295aa1616ff56da1ddf2c0ba1ac}
\examples{
data(beta10_3)
}
\keyword{datasets}
