\name{chi10}
\alias{chi10}
\docType{data}
\title{A randomly generated data}
\description{Chi10 is a randomly generated data from Chisquare
distribution with 10 degrees of freedom.}
\usage{data(chi10)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\references{http://www.sciencedirect.com/science/article/B6V8V-4NHM520-1/2/a21e0295aa1616ff56da1ddf2c0ba1ac}
\examples{
data(chi10)
}
\keyword{datasets}
