\name{iso2004}
\alias{iso2004}
\docType{data}
\title{Net sales data of 487 Turkish manufacturing firms }
\description{Net sales data of 487 Turkish manufacturing firms from the first 500 largest corporations belonging to the Istanbul Chamber of Industry
ICI in year 2004}
\usage{data(iso2004)}
\format{
  A data frame with 487 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\source{
http://www.iso.org.tr/tr/web/statiksayfalar/index.aspx?ref=3
}
\references{http://www.sciencedirect.com/science/article/B6V8V-4NHM520-1/2/a21e0295aa1616ff56da1ddf2c0ba1ac}
\examples{
data(iso2004)
}
\keyword{datasets}
