\name{normal100_10}
\alias{normal100_10}
\docType{data}
\title{A randomly generated data}
\description{normal100_10is a randomly generated data from normal
distribution with mean 100 and standart deviation 10.}
\usage{data(normal100_10)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\references{http://www.sciencedirect.com/science/article/B6V8V-4NHM520-1/2/a21e0295aa1616ff56da1ddf2c0ba1ac}
\examples{
data(normal100_10)
}
\keyword{datasets}
