\name{P75}
\alias{P75}
\docType{data}
\title{Population data of Swedish municipalities}
\description{P75 is the 1975 population (in thousands) of 284 Swedish municipalities obtained from the paper of Hedlin (2003).}
\usage{data(P75)}
\format{
  A data frame with 1000 observations on the following variable.
  \describe{
    \item{\code{V1}}{a numeric vector}
  }
}
\source{http://lib.stat.cmu.edu/datasets/mu284.txt}
\references{Hedlin, Dan (2003) Minimum variance stratification of a finite population. Southampton, UK, Southampton Statistical Sciences Research Institute, 30pp. (S3RI Methodology Working Papers, M03/07) : http://eprints.soton.ac.uk/7796/}
\examples{
data(P75)
}
\keyword{datasets}
